requirements.txt
argon2-cffi
babel
bleach
configobj
cryptography
email-validator
flask
flask-alembic
flask-babel
flask-caching
flask-login
flask-mail
flask-reuploaded
flask-session
flask-sqlalchemy
flask_weasyprint
flask-wtf
markdown
mysql-connector-python
pg8000
python-ulid
sqlalchemy
redis
waitress
weasyprint
wtforms
pytest
